+++ packages/data-loader/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.7.0](https://github.com/kintone/js-sdk/compare/@kintone/data-loader@0.6.0...@kintone/data-loader@0.7.0) (2022-02-25)
+
+
+### Bug Fixes
+
+* **data-loader:** export csv correctly when Table value is empty array ([#1349](https://github.com/kintone/js-sdk/issues/1349)) ([0f1371e](https://github.com/kintone/js-sdk/commit/0f1371e54ecdf64b57dba5d3c9307e1b4a6923a1))
+
+
+### Features
+
+* **data-loader:** support User/Department/Group Selection Field ([#1345](https://github.com/kintone/js-sdk/issues/1345)) ([7ebe9c7](https://github.com/kintone/js-sdk/commit/7ebe9c7b1aa1852569f125736b3e646e4d9e79ad)), closes [#1332](https://github.com/kintone/js-sdk/issues/1332) [#1335](https://github.com/kintone/js-sdk/issues/1335)
+
+
+
+
+
+# [0.6.0](https://github.com/kintone/js-sdk/compare/@kintone/data-loader@0.5.34...@kintone/data-loader@0.6.0) (2022-02-18)
+
+
+### Features
+
+* **data-loader:** support Attachments fields in CSV export ([#1347](https://github.com/kintone/js-sdk/issues/1347)) ([43aa526](https://github.com/kintone/js-sdk/commit/43aa526e7267aaf77102f1031b2ea9724a850173)), closes [#1332](https://github.com/kintone/js-sdk/issues/1332) [#1335](https://github.com/kintone/js-sdk/issues/1335) [#1346](https://github.com/kintone/js-sdk/issues/1346)
+
+
+
+
+
+## 0.5.34 (2022-02-14)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @kintone/rest-api-client to ^2.0.34 ([#1341](https://github.com/kintone/js-sdk/issues/1341)) ([0e01847](https://github.com/kintone/js-sdk/commit/0e018475d77c68f42d414d563377aef56a7a1d41))
+
+
+
+
+
+## 0.5.33 (2022-02-04)
+
+**Note:** Version bump only for package @kintone/data-loader
+
+
+
+
+
 ## 0.5.32 (2022-01-18)
 
 **Note:** Version bump only for package @kintone/data-loader